+++ packages/db/src/models/file.ts
@@ -3,6 +3,7 @@ import { Knex } from 'knex';
 import { BasePgModel } from './base';
 import { TableNames } from '../tables';
 
+import { convertIdFieldsToNumber } from '../lib/typeHelpers';
 import { PostgresFile, PostgresFileRecord } from '../types/file';
 
 class FilePgModel extends BasePgModel<PostgresFile, PostgresFileRecord> {
@@ -12,15 +13,16 @@ class FilePgModel extends BasePgModel<PostgresFile, PostgresFileRecord> {
     });
   }
 
-  upsert(
+  async upsert(
     knexOrTrx: Knex | Knex.Transaction,
     file: PostgresFile
   ) {
-    return knexOrTrx(this.tableName)
+    const record = await knexOrTrx(this.tableName)
       .insert(file)
       .onConflict(['bucket', 'key'])
       .merge()
       .returning('*');
+    return convertIdFieldsToNumber(record);
   }
 }