+++ packages/db/src/types/base.ts
@@ -1,4 +1,4 @@
 export interface BaseRecord {
-  cumulus_id: number,
+  cumulus_id: number | string,
   updated_at: Date,
 }